libtock: do not include syscalls
Apps should specify they use syscalls directly.

Need to duplicate the exists() function to make this work.
