+++ libtock/crypto/syscalls/aes_syscalls.h
@@ -8,7 +8,7 @@ extern "C" {
 
 #define DRIVER_NUM_AES 0x40006
 
-bool libtock_aes_exists(void);
+bool libtock_aes_driver_exists(void);
 
 returncode_t libtock_aes_set_upcall(subscribe_upcall callback, void* opaque);
