+++ libtock/crypto/syscalls/sha_syscalls.h
@@ -8,7 +8,7 @@ extern "C" {
 
 #define DRIVER_NUM_SHA 0x40005
 
-bool libtock_sha_exists(void);
+bool libtock_sha_driver_exists(void);
 
 returncode_t libtock_sha_set_upcall(subscribe_upcall callback, void* opaque);
